Fix relative path detection of ISO files
The problem becomes apparent when using symbolic links in $ISO_LOCATION
that point to ISO files on another partition.

Calling make_system_path_relative_to_its_root not on the dirname but on
the ISO file fixes this. Also avoid a double slash (//...) in $iso_path
when $rel_dirname is just the root dir of that other partition (/).